+++ src/Camino/Planner.hs
@@ -12,50 +12,84 @@ Create an optimal camino plan, based on the preferences
 -}
 
 module Camino.Planner (
-  Day,
-  Metrics(..),
-  Trip,
+    AccommodationChoice(..)
+  , AccommodationMap
+  , Day
+  , Metrics(..)
+  , Trip
 
-  accommodation,
-  daysLabel,
-  hasNonTravel,
-  nonTravelHours,
-  penance,
-  openSleeping,
-  planCamino,
-  travel,
-  travelHours,
-  tripLabel,
-  tripLegs,
-  tripStops,
-  tripWaypoints
+  , accommodation
+  , buildAccommodationMap
+  , daysLabel
+  , hasNonTravel
+  , nonTravelHours
+  , penance
+  , openSleeping
+  , planCamino
+  , travel
+  , travelHours
+  , tripLabel
+  , tripLegs
+  , tripStops
+  , tripWaypoints
 ) where
 
 import Camino.Walking
 import Camino.Camino
 import Camino.Preferences
+import Graph.Graph (available, incoming, mirror, reachable, subgraph)
 import Graph.Programming()
 import qualified Data.Map as M
 import qualified Data.Set as S
-import Data.Maybe (isJust, isNothing, fromJust, fromMaybe, mapMaybe, maybeToList)
+import Data.Maybe (isJust, fromJust, fromMaybe, mapMaybe)
 import qualified Data.Text as T
+
+data AccommodationChoice = AccommodationChoice {
+      accommodationChoice :: Accommodation -- ^ The accommodation selected
+    , accommodationChoiceServices :: S.Set Service -- ^ The additional services supplied by this accommodation
+    , accommodationChoicePenance :: Penance -- ^ The penance associated with this choice
+} deriving (Show)
+
+-- | Select accommodation based on penance / services covered / services available
+selectAccommodationChoice :: AccommodationChoice -> AccommodationChoice -> AccommodationChoice
+selectAccommodationChoice c1@(AccommodationChoice _ s1 Reject) c2@(AccommodationChoice _ s2 Reject) =
+  if S.size s1 > S.size s2 then c1 else c2
+selectAccommodationChoice (AccommodationChoice _ _ Reject) c2 = c2
+selectAccommodationChoice c1 (AccommodationChoice _ _ Reject) = c1
+selectAccommodationChoice c1@(AccommodationChoice a1 s1 p1) c2@(AccommodationChoice a2 s2 p2)
+  | p1 > p2 = c1
+  | p2 > p1 = c2
+  | S.size s1 > S.size s2 = c1
+  | S.size s2 > S.size s1 = c2
+  | S.size (accommodationServices a1) > S.size (accommodationServices a2) = c1
+  | S.size (accommodationServices a2) > S.size (accommodationServices a1) = c2
+  | otherwise = c1
+
+-- A map of accommodation choices for each location
+type AccommodationMap = M.Map Location AccommodationChoice
+-- | The trail of past/future accommodation options
+type AccommodationTail = [(AccommodationChoice, Penance)]
+-- | The tail of past/future accommodation options
+type AccommodationTrail = (AccommodationChoice, AccommodationTail)
 
 -- | The metrics for a day, segment or complete trip
 data Metrics = Metrics {
-    metricsDistance :: Float, -- ^ Actual distance in km
-    metricsTime :: Maybe Float, -- ^ Time taken in hours
-    metricsPerceivedDistance :: Maybe Float, -- ^ Perceived distance in km (Nothing if the distance is too long)
-    metricsAscent :: Float, -- ^ Ascent in metres
-    metricsDescent :: Float, -- ^ Descent in metres
-    metricsAccommodationChoice :: [Accommodation], -- ^ The places to stay on the way
-    metricsAccommodation :: Penance, -- ^ Accommodation penance in km-equivalent. This represents the distance you would be prepared to walk to avoid this accommodation
-    metricsStop :: Penance, -- ^ Stop penance in km-equivalent. The represents the costs of stopping for the night in food, urge to get on, whatever
-    metricsStopServices :: Penance, -- ^ Adjustment to penance in km-equivalnet caused by missing services at the stop point
-    metricsDayServices :: Penance, -- ^ Adjustment to penance in km-equivalnet caused by missing services throughout the day
-    metricsDistanceAdjust :: Penance, -- ^ Adjustments to distance cost in km-equivalent caused by going over/under target
-    metricsTimeAdjust :: Penance, -- ^ Adjustments to time cost in km-equivalent caused by going over/under target
-    metricsMisc :: Penance, -- ^ Additional miscellaneous penance costs and causes
-    metricsPenance :: Penance -- ^ Total penance for this leg of the metrics
+      metricsDistance :: Float -- ^ Actual distance in km
+    , metricsTime :: Maybe Float -- ^ Time taken in hours
+    , metricsPerceivedDistance :: Maybe Float -- ^ Perceived distance in km (Nothing if the distance is too long)
+    , metricsAscent :: Float -- ^ Ascent in metres
+    , metricsDescent :: Float -- ^ Descent in metres
+    , metricsAccommodationChoice :: [AccommodationChoice] -- ^ The places to stay on the way
+    , metricsAccommodation :: Penance -- ^ Accommodation penance in km-equivalent. This represents the distance you would be prepared to walk to avoid this accommodation
+    , metricsStop :: Penance -- ^ Stop penance in km-equivalent. The represents the costs of stopping for the night in food urge to get on whatever
+    , metricsMissingStopServices :: S.Set Service -- ^ The services missing at the end of the day
+    , metricsStopServices :: Penance -- ^ Adjustment to penance in km-equivalnet caused by missing services at the stop point
+    , metricsMissingDayServices :: S.Set Service -- ^ The services missing at the end of the day
+    , metricsDayServices :: Penance -- ^ Adjustment to penance in km-equivalnet caused by missing services throughout the day
+    , metricsDistanceAdjust :: Penance -- ^ Adjustments to distance cost in km-equivalent caused by going over/under target
+    , metricsTimeAdjust :: Penance -- ^ Adjustments to time cost in km-equivalent caused by going over/under target
+    , metricsMisc :: Penance -- ^ Additional miscellaneous penance costs and causes
+    , metricsPenance :: Penance -- ^ Total penance for this leg of the metrics
   } deriving (Show)
 
 instance Eq Metrics where
@@ -74,7 +108,9 @@ instance Semigroup Metrics where
     metricsAccommodationChoice = metricsAccommodationChoice m1 <> metricsAccommodationChoice m2,
     metricsAccommodation = metricsAccommodation m1 <> metricsAccommodation m2,
     metricsStop = metricsStop m1 <> metricsStop m2,
+    metricsMissingStopServices = metricsMissingStopServices m1 `S.union` metricsMissingStopServices m2,
     metricsStopServices = metricsStopServices m1 <> metricsStopServices m2,
+    metricsMissingDayServices = metricsMissingDayServices m1 `S.union` metricsMissingDayServices m2,
     metricsDayServices = metricsDayServices m1 <> metricsDayServices m2,
     metricsDistanceAdjust = metricsDistanceAdjust m1 <> metricsDistanceAdjust m2,
     metricsTimeAdjust = metricsTimeAdjust m1 <> metricsTimeAdjust m2,
@@ -83,15 +119,15 @@ instance Semigroup Metrics where
   }
 
 instance Monoid Metrics where
-  mempty = Metrics 0.0 (Just 0.0) (Just 0.0) 0.0 0.0 [] mempty mempty mempty mempty mempty mempty mempty mempty
+  mempty = Metrics 0.0 (Just 0.0) (Just 0.0) 0.0 0.0 [] mempty mempty S.empty mempty S.empty mempty mempty mempty mempty mempty
 
 instance Score Metrics where
-  invalid = Metrics 0.0 (Just 0.0) (Just 0.0) 0.0 0.0 [] mempty mempty mempty mempty mempty mempty Reject Reject
+  invalid = Metrics 0.0 (Just 0.0) (Just 0.0) 0.0 0.0 [] mempty mempty S.empty mempty S.empty mempty mempty mempty Reject Reject
 
 -- | A day's stage
 type Day = Chain Location Leg Metrics
 
--- | A complete camino
+-- | A complete camino and accommodation map
 type Trip = Chain Location Day Metrics
 
 -- Is this the last day's travel?
@@ -186,12 +222,12 @@ missingStopServices preferences _camino day = (M.keysSet $ preferenceStopService
 -- | Work out what services are missing from the desired day list
 missingDayServices :: TravelPreferences -- ^ The calculation preferences
   -> Camino -- ^ The camino model
-  -> Maybe Accommodation -- ^ The chosen accommodation
+  -> AccommodationChoice -- ^ The chosen accommodation
   -> [Leg] -- ^ The sequence of legs to use
   -> S.Set Service -- ^ The list of desired but missing services from the stop location
 missingDayServices preferences _camino accom day = let
     desired = M.keysSet $ preferenceDayServices preferences
-    desired' = desired `S.difference` (maybe S.empty accommodationServices accom)
+    desired' = desired `S.difference` accommodationChoiceServices accom
   in
     foldl (S.difference) desired' $ map locationServices (dayLocations day)
 
@@ -199,51 +235,33 @@ missingDayServices preferences _camino accom day = let
 openSleeping :: Accommodation
 openSleeping = GenericAccommodation Camping
 
-accommodation'' :: (M.Map AccommodationType Penance) -> (M.Map Service Penance) -> S.Set Service -> Accommodation -> (Maybe Accommodation, S.Set Service, Penance)
-accommodation'' accPrefs servPrefs services accom = let
+-- | Invalid accommodation choice
+invalidAccommodation :: AccommodationChoice
+invalidAccommodation = AccommodationChoice openSleeping S.empty Reject
+
+-- Select the most favourable
+accommodation' :: (M.Map AccommodationType Penance) -> S.Set Service -> Accommodation -> AccommodationChoice
+accommodation' accPrefs services accom = let
     base = M.findWithDefault Reject (accommodationType accom) accPrefs
-    missing = services `S.difference` (accommodationServices accom)
-    missingPref s = M.findWithDefault mempty s servPrefs
-    mprefs = map missingPref (S.toList missing)
+    covered = services `S.intersection` (accommodationServices accom)
   in
-    (Just accom, missing, foldl (<>) base mprefs)
+    AccommodationChoice accom covered base
     
-accommodation' :: TravelPreferences -- ^ The calculation preferences
-  -> S.Set Service -- ^ Any services not provided by the location
-  -> Location -- ^ The location to stay at
-  -> (Maybe Accommodation, S.Set Service, Penance) -- ^ Nothing if nothing is found, otherwise the penance value for the accommodation
-accommodation' preferences services location =
-  let
-    ap = preferenceAccommodation preferences
-    sp = preferenceStopServices preferences
-    accommodationOptions = if null ao then [openSleeping] else ao where ao = locationAccommodation location
-    lp = map (accommodation'' ap sp services) accommodationOptions
-    up = filter (\(_, _, p) -> p /= Reject) lp
-  in
-    foldl (\t1@(_, _, p1) -> \t2@(_, _, p2) -> if p1 < p2 then t1 else t2) (Nothing, services, Reject) up
-
--- | Calculate the accommodation penance, based on preferences
---   If there are missing services at the stop, then the suggested accommodation will be tested to see 
---   whether it offers the required services
 accommodation :: TravelPreferences -- ^ The calculation preferences
-  -> Camino -- ^ The camino model
-  -> [Leg] -- ^ The sequence of legs to use
-  -> S.Set Service -- ^ The services that, ideally, should be provided by the accommodation
-  -> Bool -- ^ Don't worry about missing accommodation
-  -> (Maybe Accommodation, S.Set Service, Penance) -- ^ Either nothing for no suitable accommodation or the resulting penance
-accommodation preferences _camino day services allowNone = 
-  let
-    acc@(accom, missing, _penance) = accommodation' preferences services (legTo $ last day)
+  -> Location -- ^ The sequence of legs
+  -> AccommodationChoice -- ^ The chosen accommodation for the location
+accommodation preferences location = let
+    services = (M.keysSet $ M.filter (/= mempty) (preferenceStopServices preferences)) `S.difference` locationServices location
+    ap = preferenceAccommodation preferences
+    accommodationOptions = if null ao then [openSleeping] else ao where ao = locationAccommodation location
+    lp = map (accommodation' ap services) accommodationOptions
+    choice = foldl selectAccommodationChoice invalidAccommodation lp
   in
-    if isNothing accom && allowNone then
-      (Nothing, missing, Penance 10.0)
-    else
-      acc
-
+    choice
 
 -- Does this leg have no acceptable accommodation (except possibly at the start and end)
-isAccommodationFree :: TravelPreferences -> [Leg] -> Bool
-isAccommodationFree preferences day = all (\l -> let (acc, _, _) = accommodation' preferences S.empty (legFrom l) in isNothing acc) (tail day)
+isAccommodationFree :: TravelPreferences -> AccommodationMap -> [Leg] -> Bool
+isAccommodationFree _preferences accommodationMap day = all (\l -> (accommodationChoicePenance $ M.findWithDefault invalidAccommodation (legFrom l) accommodationMap) == Reject) (tail day)
 
 missingServicePenance :: M.Map Service Penance -> S.Set Service -> Penance
 missingServicePenance prefs services = mconcat $ map (\s -> M.findWithDefault mempty s prefs) $ S.toList services
@@ -257,39 +275,43 @@ adjustment range bscale rscale value
 -- | Calculate the total penance implicit in a sequence of legs
 penance :: TravelPreferences -- ^ The travel preferences
   -> CaminoPreferences -- ^ The camino travelled
+  -> AccommodationMap -- ^ The map of accommodation choices
   -> [Leg] -- ^ The sequence of legs
   -> Metrics -- ^ The penance value
-penance preferences camino day =
+penance preferences camino accommodationMap day =
   let
     (_normalSpeed, _actualSpeed, time, distance, perceived, ascent, descent, nonTravel) = travelMetrics preferences day
     -- If there is no accommodation within this leg, then accept any distance. If not travelling or the last day, then skip lower bounds
     atEnd = isLastDay (preferenceFinish camino) day
     walkingSpeed = nominalSpeed Walking Normal
-    accommodationFree = isAccommodationFree preferences day
+    accommodationFree = isAccommodationFree preferences accommodationMap day
     rangeFilter = (if accommodationFree then withoutMaximum else id) . (if atEnd then withoutMinimum else id) . (if nonTravel then withoutLower else id)
     timePreferences = rangeFilter $ preferenceTime preferences
     distancePreferences = rangeFilter $ preferenceDistance preferences
     timeAdjust = maybe Reject (adjustment timePreferences 0.0 walkingSpeed) time
     distanceAdjust = adjustment distancePreferences 0.0 walkingSpeed distance
     stopMissing = missingStopServices preferences (preferenceCamino camino) day
-    (accom, stopMissing', accommodationAdjust) = accommodation preferences (preferenceCamino camino) day stopMissing atEnd -- preferred accommodation penance
+    accom = M.findWithDefault invalidAccommodation (legTo $ last day) accommodationMap -- preferred accommodation penance
+    stopMissing' = stopMissing `S.difference` accommodationChoiceServices accom
     stopMissingCost = missingServicePenance (preferenceStopServices preferences) stopMissing'
     dayMissing = missingDayServices preferences (preferenceCamino camino) accom day
     dayMissingCost = missingServicePenance (preferenceStopServices preferences) dayMissing
     stopCost = preferenceStop preferences
     distanceCost = maybe Reject Penance perceived
     misc = travelAdditional preferences day
-    total = distanceCost <> accommodationAdjust <> stopCost <> distanceAdjust <> timeAdjust <> stopMissingCost <> dayMissingCost <> misc
+    total = distanceCost <> (accommodationChoicePenance accom) <> stopCost <> distanceAdjust <> timeAdjust <> stopMissingCost <> dayMissingCost <> misc
     metrics = Metrics
       distance
       time
       perceived
       ascent
       descent
-      (maybeToList accom)
-      accommodationAdjust
+      (if accommodationChoicePenance accom == Reject then [] else [accom])
+      (accommodationChoicePenance accom)
       stopCost
+      stopMissing'
       stopMissingCost
+      dayMissing
       dayMissingCost
       distanceAdjust
       timeAdjust
@@ -300,13 +322,13 @@ penance preferences camino day =
     metrics
 
 -- | Accept a day's stage as a possibility
---   Acceptable if the time taken or distance travelled is not beyond the hard limits and allo locations are part of the selected routes.
+--   Acceptable if the time taken or distance travelled is not beyond the hard limits and all locations are part of the selected routes.
 --   Or if there is no acceptable accommodation available.
-dayAccept :: TravelPreferences -> CaminoPreferences -> [Leg] -> Bool
-dayAccept preferences camino day =
+dayAccept :: TravelPreferences -> CaminoPreferences -> AccommodationMap -> [Leg] -> Bool
+dayAccept preferences camino accommodationMap day =
   let
     atEnd = isLastDay (preferenceFinish camino) day
-    accommodationFree = isAccommodationFree preferences day
+    accommodationFree = isAccommodationFree preferences accommodationMap day
     time = travelHours preferences day
     distance = travel preferences day
     inside = isJust time && isInsideMaximum (preferenceDistance preferences) distance && isInsideMaximum (preferenceTime preferences) (fromJust time)
@@ -317,7 +339,7 @@ dayAccept preferences camino day =
 
 -- | Evaluate a day's stage for penance
 --   Acceptable if the time taken or distance travelled is not beyond the hard limits
-dayEvaluate :: TravelPreferences -> CaminoPreferences -> [Leg] -> Metrics
+dayEvaluate :: TravelPreferences -> CaminoPreferences -> AccommodationMap -> [Leg] -> Metrics
 dayEvaluate = penance
 
 -- | Choose a day's stage based on minimum penance
@@ -381,22 +403,72 @@ caminoChoice _preferences camino trip1 trip2 =
     -- trace ("Choice between " ++ (T.unpack $ tripLabel trip1) ++ " and " ++ (T.unpack $ tripLabel trip2) ++ " is " ++ (T.unpack $ tripLabel selection)) selection
     selection
 
+rundownTail :: M.Map Location AccommodationTrail -> Leg -> AccommodationTail
+rundownTail trail leg = let
+    feed = maybe [] snd (M.lookup (legFrom leg) trail)
+    distance = Penance (legDistance leg)
+    feed' = filter (\(_c, p) -> p /= mempty) $ map (\(c, p) -> (c, subtractFloor p distance)) feed
+  in
+    feed'
+
+buildAccommodationMap'' :: TravelPreferences -> Camino -> M.Map Location AccommodationTrail -> Location -> AccommodationTrail
+buildAccommodationMap'' preferences camino trail location = let
+    legs = incoming camino location
+    feed = concat $ map (rundownTail trail) legs
+    best = accommodation preferences location -- ^ The best local accommodation
+  in
+    if accommodationChoicePenance best == Reject
+    then
+      (best, feed)
+    else let
+        local = accommodationChoicePenance best
+        feed' = (best, local):feed
+        score' = maximum (map (\t -> subtractFloor (snd t) local) feed')
+      in
+        (best { accommodationChoicePenance = score' }, feed')
+
+
+buildAccommodationMap' :: TravelPreferences -> Camino -> S.Set Location -> S.Set Location -> S.Set Location -> M.Map Location AccommodationTrail -> AccommodationMap
+buildAccommodationMap' preferences camino reachable' visited next current = if S.null next
+  then
+    M.map (\(c, _t) -> c) current
+  else
+    buildAccommodationMap' preferences camino reachable' visited' next' (current `M.union` update)
+  where
+    update = M.fromSet (buildAccommodationMap'' preferences camino current) next
+    visited' = visited `S.union` M.keysSet update
+    next' = available camino reachable' visited'
+
+-- Build a forward map of accommodation choices and a backward map of choices and then choose the
+-- entry with the highest penance (meaning that there's a better option either before or after this spot)
+buildAccommodationMap :: TravelPreferences -> Camino -> Location -> Location -> (Location -> Bool) -> AccommodationMap
+buildAccommodationMap preferences camino begin end select = let
+    reachable' = reachable camino begin end select
+    sgf = subgraph camino reachable'
+    sgb = mirror sgf
+    acf = buildAccommodationMap' preferences sgf reachable' S.empty (S.singleton begin) M.empty
+    acb = buildAccommodationMap' preferences sgb reachable' S.empty (S.singleton end) M.empty
+  in
+    M.unionWith (\f -> \b -> if accommodationChoicePenance f > accommodationChoicePenance b then f else b) acf acb
+
 -- | Plan a camino based on the stated preferences
-planCamino :: TravelPreferences -> CaminoPreferences -> Either Location Trip
-planCamino preferences camino  = 
-  program 
-    (preferenceCamino camino)
-    (caminoChoice preferences camino)
-    (caminoAccept preferences camino)
-    (caminoEvaluate preferences camino)
-    (dayChoice preferences camino)
-    (dayAccept preferences camino)
-    (dayEvaluate preferences camino)
-    (\l -> S.member l allowed)
-    (preferenceStart camino)
-    (preferenceFinish camino)
+planCamino :: TravelPreferences -> CaminoPreferences -> Either Location (Trip, AccommodationMap)
+planCamino preferences camino  = either Left (\t -> Right (t, accommodationMap)) trip
   where
     allowed = allowedLocations camino
+    select l = S.member l allowed
+    accommodationMap = buildAccommodationMap preferences (preferenceCamino camino) (preferenceStart camino) (preferenceFinish camino) select
+    trip = program
+      (preferenceCamino camino)
+      (caminoChoice preferences camino)
+      (caminoAccept preferences camino)
+      (caminoEvaluate preferences camino)
+      (dayChoice preferences camino)
+      (dayAccept preferences camino accommodationMap)
+      (dayEvaluate preferences camino accommodationMap)
+      select
+      (preferenceStart camino)
+      (preferenceFinish camino)
 
 -- | Get all the stops (ie start and finish locations) on a trip in order
 tripStops :: Trip -> [Location]