+++ src/Camino/Server/Forms.hs
@@ -439,7 +439,7 @@ chooseServicesForm help prefs extra = do
     let accommodationOptions = map (\v -> (v, [ihamlet|^{caminoAccommodationTypeIcon v}&nbsp;_{caminoAccommodationTypeMsg v}|] messages router)) accommodationTypeEnumeration
     let stopServiceOptions = map (\v -> (v, [ihamlet|^{caminoServiceIcon v}&nbsp;_{caminoServiceMsg v}|] messages router)) serviceEnumeration
     let dayServiceOptions = map (\v -> (v, [ihamlet|^{caminoServiceIcon v}&nbsp;_{caminoServiceMsg v}|] messages router)) townServiceEnumeration
-    (stRes, stView) <- mreq penanceField (fieldSettingsLabel MsgStopPreferencesLabel) (prefStop <$> prefs)
+    (stRes, stView) <- mreq penanceField (fieldSettingsLabelTooltip MsgStopPreferencesLabel MsgStopPreferencesText) (prefStop <$> prefs)
     (acRes, acView) <- mreq (penanceMapField True accommodationOptions) (fieldSettingsLabel MsgAccommodationPreferencesLabel) (prefAccommodation <$> prefs)
     (ssRes, ssView) <- mreq (penanceMapField False stopServiceOptions) (fieldSettingsLabel MsgStopServicePreferencesLabel) (prefStopServices <$> prefs)
     (dsRes, dsView) <- mreq (penanceMapField False dayServiceOptions) (fieldSettingsLabel MsgDayServicePreferencesLabel) (prefDayServices <$> prefs)
@@ -462,6 +462,9 @@ chooseServicesForm help prefs extra = do
           <div .col>
             <label for="#{fvId view}">
               ^{fvLabel view} ^{help}
+            $maybe tt <- fvTooltip view
+              <div .form-text>
+                ^{tt}
             ^{fvInput view}
       $with view <- viewAccommodation fields
         <div .row .mb-3>